+++ src/test/java/fr/marcwrobel/jbanking/AgreementTest.java
@@ -1,5 +1,10 @@
 package fr.marcwrobel.jbanking;
 
+import static fr.marcwrobel.jbanking.Agreement.EUROPEAN_ECONOMIC_AREA;
+import static fr.marcwrobel.jbanking.Agreement.EUROPEAN_FREE_TRADE_ASSOCIATION;
+import static fr.marcwrobel.jbanking.Agreement.EUROPEAN_UNION;
+import static fr.marcwrobel.jbanking.Agreement.SEPA_COM_PACIFIQUE;
+import static fr.marcwrobel.jbanking.Agreement.SINGLE_EURO_PAYMENTS_AREA;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -24,15 +29,29 @@ class AgreementTest {
 
   @Test
   void sepaComPacifiqueParticipantsAreNotSingleEuroPaymentAreaParticipants() {
-    for (IsoCountry participant : Agreement.SEPA_COM_PACIFIQUE.getParticipants()) {
-      assertFalse(participant.isParticipatingTo(Agreement.SINGLE_EURO_PAYMENTS_AREA));
+    for (IsoCountry participant : SEPA_COM_PACIFIQUE.getParticipants()) {
+      assertFalse(participant.isParticipatingTo(SINGLE_EURO_PAYMENTS_AREA), participant::name);
     }
   }
 
   @Test
   void europeanUnionParticipantsAreAlsoSingleEuroPaymentAreaParticipants() {
-    for (IsoCountry participant : Agreement.EUROPEAN_UNION.getParticipants()) {
-      assertTrue(participant.isParticipatingTo(Agreement.SINGLE_EURO_PAYMENTS_AREA));
+    for (IsoCountry participant : EUROPEAN_UNION.getParticipants()) {
+      assertTrue(participant.isParticipatingTo(SINGLE_EURO_PAYMENTS_AREA), participant::name);
+    }
+  }
+
+  @Test
+  void europeanUnionParticipantsAreEuropeanEconomicAreaParticipants() {
+    for (IsoCountry participant : EUROPEAN_UNION.getParticipants()) {
+      assertTrue(participant.isParticipatingTo(EUROPEAN_ECONOMIC_AREA), participant::name);
+    }
+  }
+
+  @Test
+  void europeanFreeTradeAssociationParticipantsAreEuropeanEconomicAreaParticipants() {
+    for (IsoCountry participant : EUROPEAN_FREE_TRADE_ASSOCIATION.getParticipants()) {
+      assertTrue(participant.isParticipatingTo(EUROPEAN_ECONOMIC_AREA), participant::name);
     }
   }
 }